move Reachability related objects to new file and namespace per https://github.com/coverlet-coverage/coverlet/pull/904#issuecomment-675925575
